+++ quackosm/__init__.py
@@ -13,10 +13,11 @@ from quackosm.functions import (
     convert_pbf_to_geodataframe,
     convert_pbf_to_parquet,
 )
+from quackosm.geocode import geocode_to_geometry
 from quackosm.pbf_file_reader import PbfFileReader
 
 __app_name__ = "QuackOSM"
-__version__ = "0.8.2"
+__version__ = "0.8.3"
 
 __all__ = [
     "PbfFileReader",
@@ -25,5 +26,9 @@ __all__ = [
     "convert_osm_extract_to_parquet",
     "convert_pbf_to_geodataframe",
     "convert_geometry_to_geodataframe",
+
     "convert_osm_extract_to_geodataframe",
+
+    "geocode_to_geometry",
+
 ]